LayoutTests/imported/w3c/web-platform-tests/html/semantics/popups/popup-attribute-basic.tentative-expected.txt

Pop up
Pop up
Pop up
Pop up
Pop up
Not a pop-up
Not a pop-up
Not a pop-up
Animated pop-up

FAIL The .showPopUp() and .hidePopUp() work on a pop-up, for <div popup="" id="boolean">Pop up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() work on a pop-up, for <div popup="">Pop up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() work on a pop-up, for <div popup="auto">Pop up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() work on a pop-up, for <div popup="hint">Pop up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() work on a pop-up, for <div popup="manual">Pop up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() do NOT work on elements without a 'popup' attribute, <div>Not a pop-up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() do NOT work on elements without a 'popup' attribute, <div popup="popup">Not a pop-up</div>. The string did not match the expected pattern.
FAIL The .showPopUp() and .hidePopUp() do NOT work on elements without a 'popup' attribute, <div popup="invalid">Not a pop-up</div>. The string did not match the expected pattern.
FAIL IDL attribute reflection assert_equals: expected (string) "auto" but got (undefined) undefined
FAIL Popup attribute value should be case insensitive The string did not match the expected pattern.
FAIL Changing attribute values for pop-up should work The string did not match the expected pattern.
FAIL Changing attribute values should close open pop-ups popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL Removing a visible popup=auto element from the document should close the pop-up popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL A showing popup=auto does not match :modal popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL Removing a visible popup=hint element from the document should close the pop-up popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL A showing popup=hint does not match :modal popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL Removing a visible popup=manual element from the document should close the pop-up popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL A showing popup=manual does not match :modal popUp.showPopUp is not a function. (In 'popUp.showPopUp()', 'popUp.showPopUp' is undefined)
FAIL Exceptions are thrown even when show/hide are animated assert_false: expected false got true
